use central maven dependencies in poms
this removes the dependency on our custom rxtx and custom jna.

the rxtx dependency will still be needed for the best experience
at runtime, and the custom binaries are still in the sensor/bin folder

the jna dependency was to help out on webstart, so this can be added 
in the webstart application pom, by excluding the central jna
and adding our custom one.