treat all resource urls the same
This was treating file: resource urls differently.
It would save a file copy, but this means a production code path
would not be used in a typical development environment.
It also caused a bug when running tests because the library file
would be deleted after it was loaded and in the file: case this
deleted the original file so future tests couldn't find it.