Better handling of Go Direct sensors that don't return values
Go Direct motion sensor is showing two channels that do not return values in testing.  Improve handling of channels that do not return values and ensure that we return the correct number of measurements when reading an experiment (should be bound by the valid channel with the min number of measurements greater than 0).
[#161353930]